Added Variable jump and a passthough ability.
Added variable jump (the longer space is held the higher player jumps(max and min jump values added)) Passthough tag allows for platfroms to be passed though by jumping under them or pressing down when on them. This ability is togolable in the player script and can be used when poo baby is in his gas form.