BUG FIX: Slowing when switching state on slope & Add TODO for testing
I believe the problem was the sprite being sort of inside the wall, so i put a tiny increase in height to avoid any "smushing" of the wall/rays.

A better fix would likely be in the Controller, but that's Daniel's domain.

TODO: MAY NOT WORK ON ALL SLOPES, FURTHER TESTING NEEDED!